requirements.txt
streamlit>=1.35
google-genai>=1.16
opencv-python-headless>=4.9
yt-dlp>=2025.1
requests>=2.31
plotly>=5.22
pandas>=2.1
numpy>=1.26
Pillow>=10.2
ollama>=0.4
python-dotenv>=1.0
reportlab>=4.1
kaleido>=0.2
